v1.8.1
This is a patch release with some bug fixes and performance improvements.

There is a bug in lu v0.3.0 with regards to how it handles serialisation
of the configuration file. If you get errors about missing entries that
are actually present but commented out, make sure to `dub upgrade`.
Not bumping the dependency version since it technically still works with
v0.3.0 with only minor deficiencies.

Other changes:

* Updated arsd-official dependency to v7.2.0
* Fixed Oneliners not getting a configuration file entry
* Fixed a bug where we weren't passing a connection timeout duration,
  making lu v0.3.0 time out after 600_000 seconds instead of 600.
  (It is now also fixed in upstream lu v0.3.1)
* Removed last use of regex in favour of own slimmed-down state machines!
  The only time there will be regex is if a plugin implements a
  `BotRegex` command, which none currently do.
* Webtitles now caches YouTube lookups too
* The previously global `kameloso.common.settings` and
  `kameloso.common.logger` are now thread-local. Threads beware.
* Behind-the-scenes refactoring, documentation, modules
* Various performance improvements

Feedback appreciated.