Revisit onEventImpl ordering again
This approach seems to be the best of all worlds. Check UDAs first and
set a runtime flag if one matched, then break out of the UDA loop and
either continue (based on the value of the match bool) or return
Next.continue_.

The amount of extra code generated is minimised; a debug -c dev is down
to ~4464 Mb.