Rework method of IRCPlugin initialisation
So far we've been keeping an AliasSeq of IRCPlugin subclasses in
`kameloso.plugins`, which we import and foreach over, instantiating
plugins as we go. This had several drawbacks; for one, we had to jump
through hoops to be able to allow for easy disabling of plugins, with
some annoying and buggy import logic.

This approach keeps an AliasSeq of plugin modules, which are then
foreached over to find IRCPlugin subclasses to instantiate. We can then
just introspect the module to see if there are any such classes, and
skip it if so, without having to skip importing and faking symbols.

It's just better. Let's see if it works too.