v1.9.1
There were too many changes since v1.9.0, so releasing this instead of
v2.0.0 as we wait for things to settle down and for new bugs to surface.

Many plugins had obscure configurable options removed to combat config
creep. They can still be manually set with `--set plugin.option=value`,
such as with a script.

* Compilation memory requirements up slightly but not significantly;
  Linux debug application build is now ~3340 Mb
* getopt touch-up, now shows some current values
* Added Stopwatch plugin; `!stopwatch` `[start|status|stop]`
* Added Counter plugin (`!counter`); used to count the occurrence of
  something, like `!deaths+` when streaming a game
* Added Twitch bot command `!shoutout` (alias `!so`)
* Improved other Twitch bot commands; for example, now `!uptime` and
  `!stop` report viewer statitics
* Miscellaneous renames, moves and refactorings to combat
  `somepackage.common` creep

Hopefully v2.0.0 is next, but there might be a v1.9.2.