v3.0.0-beta.2
A lot of changes, but we're slowly converging on v3.0.0. Still a few
betas away, I think.

Compilation memory required has been lowered noticeably since beta.1;
mostly from being smarter and breaking up large variadic templates (such
as `formatObjects`) into one general template which in turn instantiates
smaller templates based on the variadics. This meant that instantiations
could be reused where the same kind of object appeared in different
instantiations of the general template, and memory usage dropped.

    someTemplate(string, int, bool);
    someTemplate(int, string, bool);

This instantiates two different `someTemplate` templates, simply because
the variadics differ, even though they contain the same types. Splitting
`someTemplate` up into `someTemplateStringImpl`, `someTemplateIntImpl`
and `someTemplateBoolImpl` meant that only the actually "dispatching"
bit of `someTemplate` need be unique per instantiation. The types in the
actual variadic could now be reused.

`onEventImpl` was refactored somewhat and now passes the
`IRCEventHandler` UDA object as a runtime value. It's arguably less
elegant in that less is being done at compile-time, but memory use
dropped further.

Support for older compilers was added back, but in a way where a
ridiculous amount of array literals were being generated, several for
each event. This was not a problem when we limited ourselves to 2.096
and later, but the easy fix to get support for older versions back
introduced it. Partially solved; the sole known side-effect is that
compilers pre-2.096 cannot mark an event handler function as
`.verbose(true)`. (The entire module must be made verbose by way of
passing `Yes.debug` to the awareness mixins.) Garbage collector pressure
is back at acceptable levels after this change.

"Repeat" vs "Replay" continues to be a source of confusion. For the time
being, "Repeats" were renamed "Reparses", because that's what they do.
Ideally there would be a way to combine the two, but that's not a
priority right now. It can be done behind the scenes as a non-breaking
change. Lastly, reparsing was changed from being a Fiber solution to a
simple delegate.

Needs testing.