v3.0.0-rc.1
As before, mostly plumbing invisible to users.

Notably the garbage collector was given some initial values, with regards
to initial size and pool growth increments. It should lead to fewer
collections over time at the expense of slightly higher memory consumption
at program start, but it's in the order of the odd megabytes. Users can
override this with the `--DRT-gcopt=` getopt flag if it somehow turns out
to be harmful.

* Users should see the pressure on the garbage collector go down
  significantly when the Seen plugin is enabled
* Developers should not have to go through the considerable mental
  gymnastics previously needed to colour terminal messages anymore
* Logging to file was broken and fixed
* ldc works again with an extra compiler flag in `DFLAGS`, though the real
  fix (increased stack size) is confirmed to be due in 1.29.0
* Version `TwitchAPIFeatures` was hardcoded as enabled and subsequently
  removed

Maybe one more release candidate after this, then v3.0.0 final.