Refactor longest member/type name traits
Previously:
    kameloso.traits.longestMemberName
    kameloso.traits.longestUnserialisableMemberName
    kameloso.traits.longestMemberTypeName
    kameloso.traits.longestUnserialisableMemberTypeName

Now:
    kameloso.traits.longestMemberNames
    kameloso.traits.longestUnserialisableMemberNames

Before the templates aliased themselves to the longest member or member
type name of the passed aggregates respectively; now they alias themselves
to a Voldemort struct containing both the names of the longest member and
the name of the longest type.

`printing.d` was updated to reflect this.