v3.0.0-rc.2
Likely final release candidate.

Outgoing IRC messages now support tags the way terminal logger messages
do. This technically violated feature freeze. I'm not super happy about
the syntax and it may change in future versions, but for now it works
and greatly simplifies highlighting in messages.

Several configuration entries were deemed too fringe and made hidden.
They're still available via `--set plugin.option=value`, or as a new
version identifier in the case of `TwitchBotSettings.promoteEverywhere`
(version `TwitchPromoteEverywhere`).

`ConnectSettings.sendAfterConnect` was changed from a `string[]` to a
`string`, simply because of how getopt would split entries in it by
commas (which we want it to do to other flags). Provided the `;;`
delimiter was used, users should not need to change anything in the
configuration file.