v3.0.0-rc.3
So I was wrong and we need a third release candidate.

Most changes since last relate to the initial impression on Windows. We
had to move from `std.net.curl` back to `dlang-requests` since Windows
doesn't come with a `libcurl.dll`. We'd have to require users to install
a D compiler first (or something similar that includes cURL), which would
not be ideal.

Then it became apparent the handling of private key and certificate files
were less stellar than they could be. Much later and now it's as smooth as
we can get it without shipping OpenSSL libs and cURL's `cacert.pem`
certificate bundle ourselves.

Now a user should just have to install OpenSSL (or use an alternative
shell like MinGW64, Cygwin, ...), potentially grab a `cacert.pem` and
they're set. Some error messages now include links to our wiki page on
OpenSSL too, to make it even easier.

There should not be a need for a fourth release candidate.