v3.0.0-rc.5
One truly final release candidate after the migration to `arsd.http2`.

You can now download and launch the installer for OpenSSL for Windows by
passing `--get-openssl`. This should be enough to be able to connect to
SSL servers and have plugins access the web via https. If a certificate
bundle is required for some reason, `--get-cacert` now downloads one for
you, saves it in the configuration directory and sets the configuration
file to point to it.

There should really really not be a need for a sixth.