v3.0.0
Release!

Much has been done in the plumbing since `v2.x`. The way the bot
*behaves* has not changed, but the way it *works* saw some spanning
refactors. The end experience should not differ much from previous
versions (aside from general improvements and bugs fixed), but quality
of life for developers should be significantly improved.

The way event handler functions are annotated to be automatically called
on incoming events was changed in a non-backwards-compatible way, and
custom plugins will have to have their annotations rewritten to match
the new scheme. It can be automated to a fairly high degree though;
inquire within.

A considerable amount of effort was spent improving initial setup on
Windows. Out of the box, the bot cannot make encrypted connections to
servers, nor access the web via secure `https://` addresses. For this a
third-party OpenSSL library has to be installed, which is
regrettable but at this point a fact of life. To facilitate installing
this there's now a `--get-openssl` command-line flag you can pass, which
makes it automatically download and launch the installer for OpenSSL
for Windows. Once in place it should just work.

On the horizon:

* adding support for multiple syntax entries in event handler help
  descriptions
* breaking out Twitch timers into its own, generic Timers plugin (since
  there's nothing really Twitch-specific about it)
* exploring more Twitch commands (`ecount`, `settitle`, `setgame`,
  `vanish`/`poof`, `watchtime`, `repeat`/`spam`?)
* experimenting with setting up a homebrew web server to smoothen out
  Twitch key generation
* ideas needed. *testing* needed

Lastly, the project is now licensed under the Boost Software License,
like much of the D ecosystem.

Please report bugs.