v3.2.0
This release contains additions and fixes across the board and as such
bumps the minor version.

New to Twitch is song requests. The idea is that using `!songrequest`,
one can add a link to a YouTube or Spotify playlist, which the streamer
then keeps playing in the background. This requires Google/YouTube
and/or Spotify authorisation tokens (and playlist IDs), and new keygens
(`--set twitch.googleKeygen`, `--set twitch.spotifyKeygen`) were added
to facilitate entering such.

More Twitch command additions in the form of `!setgame`, `!settitle` and
`!commercial`. As their names suggest, these allow you to set the
current game and channel title, as well as playing commercials. Notably
different from other commands though is that these require authorisation
tokens with permissions of the channel owner, so yet another keygen was
added for this (`--set twitch.superKeygen`).

Experimental support landed for Twitch polls. Help from a Twitch
affiliate is needed to verify that it's not flat-out broken, but
technically it should work.

The Votes plugin saw some refactoring to make it easier to mentally
follow. It also gained configuration settings to limit valid votes to
people online, and to forbid vote choices starting with the command
prefix.

Durations may now be supplied as abbreviated `*d*h*m*s` strings, instead
of only as an amount in seconds.

You now get occasional reminders as your Twitch authorisation key is
about to expire.

Added Twitch command `!nuke`. This retroactively deletes recent messages
containing a word or phrase.

Oneliners response lines can now be edited in-place.

There's more, but that's largely it. Please report bugs.