twitch: Pad reminder expiry time by 59 seconds
When the duration is formatted the time is floored. A timer at 6h0m
remaining is thus almost guaranteed to report a duration of 5h59m.

To combat this, we padded the expiry duration by 5 seconds, to make sure
it is always floored to the correct minute. There is however no reason
not to pad with the full 59 seconds for more headroom. The 6h0m duration
would thus become 6h0m59s, floored to the proper 6h0m.