v3.4.1
This is an incremental increase with mostly internal changes that should
be invisible to the user.

Many event handlers that were previously manually creating Fibers are
now annotated `.fiber(true)` and automatically gets called from within a
*new* Fiber upon triggering. This lets us add some static asserts that
`IRCEvent`s are not passed by `ref` to those functions, which is
otherwise prone to segfaults and memory corruption in release builds.

Note that subsequent invocations of the same event handler **will result
in the creation of new Fibers**, so in some cases (particularly with
`await`ing Fibers) the previous way of doing it might still be
preferable. Alternatively, care has to be taken to early-return if a
previous Fiber is still running, tracked by some plugin member or
static/global variable.

The Twitch plugin now fetches a live list of the top online bots (from
https://twitchinsights.net) every chatters update for exclusion from the
viewer count, instead of having a manually defined blacklist of that
quickly grows out of sync.

* Added `TwitchSettings.watchtimeExcludesLurkers`; if true, makes
  `!watchtime` only take into account viewers who have shown activity
  this broadcast.
* Added `PrinterSettings.logPrivateMessages`

Signed-off-by: JR <zorael@gmail.com>