v3.4.2
This is a minor patch that merely remedies some egregious allocations.

Chatbot's `DANCE` routine, which is annotated to trigger on every
`IRCEvent.Type.CHAN` and `IRCEvent.Type.QUERY`, was recently annotated
`.fiber(true)` and had its whole body made into a function appropriate
for a Fiber. It worked and it was elegant, but the original function did
several checks before even creating the Fiber, to determine whether or
not the message matched the expression `\bDANCE[ !.?]+$`. This
opting-out code was now included in the Fiber body, and without any
preceding code to detect the presence of the triggering text and return
early, a new Fiber was instantiated on every message.

The event handler is now a normal `.fiber(false)` function again, and
creates a new Fiber only if the checks pass.

Signed-off-by: JR <zorael@gmail.com>