Add showcase toy `SamePlugin`
It replies to channel messages of "same" with "same".

It showcases two approaches; one by simple string comparisons done in
the function body (does the incoming channel message's content equal
"same"?), the other by regular expression at the annotation level.

In the first approach, the event handler function is always called on
all channel messages, and it is up to the body to decide whether the
message was interesting. In the second approach, the function is only
ever called if the expression matched, so the body can safely assume the
message was something it should react to.

It's disabled in `dub.sdl` by default. Include version "WithSamePlugin"
to enable it.