Better introspection of whether symbols exist
So far we've been using `__traits(compiles, sym)` to detect whether some
`sym` was present, but it occasionally failed when `sym` existed but
failed to compile.

This new approach tries to detect its presence by seeing if an alias of
it can be declared, with `__traits(compiles, { alias _ = sym; })`.

Hopefully it should be more robust with less false negatives.

Signed-off-by: JR <zorael@gmail.com>