Initial code dump; automatic plugin registration
This changes the way we instantiate plugins, from having a defined list
of (strings of) plugin modules that we iterate through and mix in imports
of, to instead have individual modules register themselves to be
instantiated by use of module constructors.

Some functions had to be broken out into new modules to avoid cyclic
dependencies between plugins.