Copy/paste NAMES reply user discovery into Persistence
This was originally part of user awareness, but it makes sense to have
Persistence discover users from NAMES too, without mixin in the full
user awareness.

Ideally we would be able to reuse the code, but that would mean making
the private function public. Out of the two evils, copy/pasting is
arguably the lesser one.