v3.7.0
2023!

This release includes a rewrite of the Notes and the Quotes plugins. Old
`notes.json` files should just work without manual intervention, but old
`quotes.json` files are not compatible, due to quotes historically not
being separated by channel (which they are now). They will be converted
into the new format on a best-effort basis, but manual editing with a
text editor must be done to properly finish the conversion.

The Timer plugin also saw refactoring. It is now less reliant on
closures within Fibers and thus now supports suspend and resume. It is
now also compiled in even when Twitch support is disabled. The
limitation was never meant to be and is a remnant of when timers were
part of the Twitch plugin.

The Twitch plugin now knows how to detect streams being online, and the
commands `!start` and `!stop` that were used to manually mark the
beginnings and ends of them were subsequently removed. The polling is
done once a minute, which should be good enough.

Related, the Twitch plugin will now retry failed API queries for a total
of three times before admitting failure. This is to combat spurious
empty responses from the server.

`bash.org` quotes were broken out of the Chatbot plugin and into its own
Bash plugin.

A new system of plugin registration was added. With it, plugins no
longer need to have their modules declared in `plugins/package.d` to be
included and instantiated on program startup. In its stead, each module
must now instead mix in `ModuleRegistration`, which will make the
program automatically pick their plugin classes up.

Due to this change, we can no longer support compilers prior to
**2.085** (March 2019), up from **2.084**. Apparently improvements were
made to module import cycle detection in that release, and we now rely
on the improved behaviour.

A new Time plugin was added, used to query what the time is in different
timezones. It makes more sense on Twitch but it's generic enough to have
as its own plugin.

`gdc` now works but may fail to link, depending in your particular
setup. On one Debian installation it works fine with 12.1.0, on a
Manjaro/Arch installation it fails with 12.2.0.