v3.7.1
This release includes a fix for Twitch background monitor Fibers ending
upon receiving unexpected server responses. This could happen due to
connection outages transient enough that the bot stays connected to the
server, but long enough that server queries failed. With this the
monitors now endlessly retry for as long as we stay in a channel. The
plugin also no longer overlaps multiple monitor Fiber instances
monitoring the same channel, which was a fringe edge-case that you would
run into if you left a channel and rejoined it.

The Time plugin can now express the time in 12-hour AM/PM, if set to do
so in the configuration file. The default remains 24-hour time.

Polls now survive disconnects and exits/restarts of the program. They do
not survive unrecoverable crashes though, such as due to segmentation
faults or being killed by signals. Naturally, votes placed when the bot
isn't there to hear them aren't magically counted.