Add opt-in PING monitor
This adds a monitor Fiber that sends a PING to the server if we haven't
received one for 10 minutes. This seems to happen sometimes with dead
connections that can still be read from. Sending anything at that point
seems to successfully make it detect the disconnect. So automate that.

Since it's not completely obvious if this is necessary, make it opt-in
for now.