Migrate from registering modules to registering plugins
The previous attempt was to register a plugin's module for
introspection, and then the plugin class therein would be queued for
instantiation on program startup. This incurred unnecessary
introspection, when the plugin class could just be specified at the
registration mixin site.

This replaces the mixin `ModuleRegistration` with `PluginRegistration`
that accepts just that.