v3.12.0
This release includes a rewritten Pipeline plugin that does
non-blocking reads. To the user it should behave exactly as before,
but it now no longer uses a separate thread that sits blocking on reading
from the pipe. It was both simple and needlessly complicated but as a
quick hack it sort of worked, except when it didn't. As before, this
plugin remains Posix-only.

Enabled the `-dip1000` compilation flag for scoped pointers.
This regrettably bumps the minimum dmd version from 2.085 to 2.086
(and ldc from 1.14 to 1.16) to resolve linker errors. At the time of
writing (August 2023) we're at 2.105, so 2.086 from May 2019 is still
decidedly ancient.

Disabled the `-dip25` compilation flag for sealed references.
Contemporary compilers enable it by default now, and explicitly
including it just generates "`-dip25 no longer has any effect`" noise
when compiling.

* Refactored plugin initialisation in such a way that allows Twitch
  keygens to perform their captive routines without first connecting
  to a server
* Improved Google and Spotify keygens to the point of usability
* Made throttling/scheduling of outgoing messages more precise
* Bumped minimum dialect dependency version to v2.0.4
* Bumped lu dependency version to v2.0.0
* ldc versions older than 1.25 fail to link in release mode and seem to
  have been for some time, but what do