v3.12.1
This is primarily an emergency hotfix for non-super Twitch key generation
not saving the key to the configuration file.

Additionally changed the API endpoint used for Twitch channel follower
queries. The hitherto-used `helix/users/follows` was deprecated in
February 2023 and is scheduled to be shut down on September 12th.
The new endpoint requires additional OAuth scopes, so a new key will have
to be generated for the `!followage` command to work.

* Refactored the Webtitle plugin slightly
* Now generally uses `.assumeUnique` instead of `.idup` wherever applicable
* Fixed the CircleCI build script always failing to fetch the compiler
  due bash `set -u`