twitch: Use several worker threads for HTTP queries
The number of which is configurable in TwitchSettings.workerThreads.

They are called in a round-robin fashion, which isn't ideal in the sense
that one blocking request can still lock everything up for further
queries that are sent to the same thread. It's a simple solution but
should hopefully work well enough.