Undo application of kameloso.string.replaceAA
It's still neat but the way we use it we would have to incur an
associative array literal on each call. Keeping around static AAs only
work in certain conditions, which (for our purposes) rarely apply.

This removes the string-replacement in
kameloso.configreader.configurationText; meaning, `[Votes]` is no longer
rewritten as `[Poll]`, etc.