Replace the message fiber with a normal function
The reason for making it a fiber in the first place was to avoid the GC
allocating closures for the delegates we passed to receiveTimeout.

We don't use receiveTimeout at all now unless version
WantConcurrencyMessageLoop is declared, and if it is we can avoid the
closures by use of scope variables with compilers 2.092 and later.

Remove support for ThreadMessage.MessageType.{send,quiet,immediate}line
while we're at it.